git subrepo pull --branch=vsk-cassie --force godot
subrepo:
  subdir:   "godot"
  merged:   "de12d58559"
upstream:
  origin:   "https://github.com/V-Sekai/godot.git"
  branch:   "vsk-cassie"
  commit:   "de12d58559"
git-subrepo:
  version:  "0.4.6"
  origin:   "https://github.com/ingydotnet/git-subrepo"
  commit:   "110b9eb"

+++ godot/modules/vorbis/audio_stream_ogg_vorbis.h
@@ -107,9 +107,9 @@ class AudioStreamOggVorbis : public AudioStream {
 	friend class AudioStreamPlaybackOggVorbis;
 
 	int channels = 1;
-	double length = 0.0;
+	float length = 0.0;
 	bool loop = false;
-	double loop_offset = 0.0;
+	float loop_offset = 0.0;
 
 	// Performs a seek to the beginning of the stream, should not be called during playback!
 	// Also causes allocation and deallocation.
